+++ Mathematics_Algo/modInverse.cpp
@@ -1,21 +1,21 @@
-int gcd(int a,int b,int &x,int &y)
+long long gcd(long long a,long long b,long long &x,long long &y)
 {
     if (b == 0) {
         x = 1;
         y = 0;
         return a;
     }
-    int x1, y1;
-    int d = gcd(b, a % b, x1, y1);
+    long long x1, y1;
+    long long d = gcd(b, a % b, x1, y1);
     x = y1;
     y = x1 - y1 * (a / b);
     return d;
 }
 
-int modInverse(int a,int m)
+long long modInverse(long long a,long long m)
 {
-    int x, y;
-    int g = gcd(a, m, x, y);
+    long long x, y;
+    long long g = gcd(a, m, x, y);
     if (g != 1)
         return -1;
     else
