+++ Mathematics_Algo/nCr_without_Mod.cpp
@@ -1,6 +1,6 @@
-ll nCr(ll n,ll r){
-    ll result = n;
-    ll i;
+long long nCr(long long n,long long r){
+    long long result = n;
+    long long i;
     rep(i,2,r+1){
         result*=(n-i+1);
         result/=i;
